requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.2
matplotlib==3.8.0
seaborn==0.12.2
lightfm==1.16
tensorflow==2.14.0
keras==2.14.0
scikit-surprise==1.1.1
jupyter==1.0.0
notebook==7.0.3
